requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
matplotlib>=3.4.0
aiohttp>=3.7.4